Pull out srl v8 generator into generator_bases
